merging the changes from 'dev' branch

+++ handlers/progress/UpdateProgressItemHandler.js
@@ -50,6 +50,10 @@ const UpdateProgressItemHandler = async (req, res) => {
 			return res.status(404).json(handleApiError({ status: 404 }));
 		}
 
+		if (String(progressItem.userId) !== String(req.user?.userId)) {
+			return res.status(403).json(handleApiError({ status: 403 }));
+		}
+
 		if (progressItem.image) {
 			deleteUploadedFile(getFilenameFromUrl(progressItem.image));
 		}
@@ -57,7 +61,7 @@ const UpdateProgressItemHandler = async (req, res) => {
 		Object.assign(progressItem, data);
 		await progressItem.save();
 
-		return res.status(200).json(progressItem);
+		return res.status(200).json({ progressItem });
 	} catch (err) {
 		return res.status(500).json(handleApiError({ err }));
 	}
